refactor(agent): isolate non-stream request polling
